+++ apps/judicial-system/api/src/app/modules/court/court.resolver.ts
@@ -1,8 +1,8 @@
 import { Inject, UseGuards } from '@nestjs/common'
 import { Args, Context, Mutation, Resolver } from '@nestjs/graphql'
 
-import type { Logger } from '@island.is/logging'
 import { LOGGER_PROVIDER } from '@island.is/logging'
+import type { Logger } from '@island.is/logging'
 import type { CaseType, User } from '@island.is/judicial-system/types'
 import {
   AuditedAction,
