+++ apps/judicial-system/backend/src/app/modules/case/filters/case.filters.spec.ts
@@ -1,6 +1,7 @@
 import { Op } from 'sequelize'
 
-import { CaseState, User, UserRole } from '@island.is/judicial-system/types'
+import { CaseState, UserRole } from '@island.is/judicial-system/types'
+import type { User } from '@island.is/judicial-system/types'
 
 import { getCasesQueryFilter, isCaseBlockedFromUser } from './case.filters'
 import { Case } from '../models'
