+++ apps/judicial-system/backend/src/app/modules/user/dto/createUser.dto.ts
@@ -2,7 +2,7 @@ import { IsBoolean, IsNotEmpty, IsString } from 'class-validator'
 
 import { ApiProperty } from '@nestjs/swagger'
 
-import { UserRole } from '@island.is/judicial-system/types'
+import type { UserRole } from '@island.is/judicial-system/types'
 
 export class CreateUserDto {
   @IsNotEmpty()
