+++ apps/judicial-system/backend/src/app/modules/user/dto/updateUser.dto.ts
@@ -2,7 +2,7 @@ import { IsBoolean, IsOptional, IsString } from 'class-validator'
 
 import { ApiProperty } from '@nestjs/swagger'
 
-import { UserRole } from '@island.is/judicial-system/types'
+import type { UserRole } from '@island.is/judicial-system/types'
 
 export class UpdateUserDto {
   @IsOptional()
