+++ apps/judicial-system/backend/src/app/modules/user/user.model.ts
@@ -11,7 +11,7 @@ import {
 
 import { ApiProperty } from '@nestjs/swagger'
 
-import { UserRole } from '@island.is/judicial-system/types'
+import type { UserRole } from '@island.is/judicial-system/types'
 
 import { Institution } from '../institution'
