+++ apps/judicial-system/backend/src/app/modules/user/user.service.ts
@@ -1,8 +1,8 @@
 import { Inject, Injectable } from '@nestjs/common'
 import { InjectModel } from '@nestjs/sequelize'
 
-import type { Logger } from '@island.is/logging'
 import { LOGGER_PROVIDER } from '@island.is/logging'
+import type { Logger } from '@island.is/logging'
 import { UserRole } from '@island.is/judicial-system/types'
 
 import { environment } from '../../../environments'
