+++ apps/judicial-system/web/src/routes/Admin/ChangeUser/ChangeUser.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react'
 
 import { PageLayout } from '@island.is/judicial-system-web/src/shared-components'
-import { User } from '@island.is/judicial-system/types'
+import type { User } from '@island.is/judicial-system/types'
 import { useMutation, useQuery } from '@apollo/client'
 import * as Constants from '@island.is/judicial-system-web/src/utils/constants'
 import {
