+++ apps/judicial-system/web/src/routes/Admin/Users/Users.tsx
@@ -2,7 +2,8 @@ import React, { useEffect } from 'react'
 import cn from 'classnames'
 import { AlertMessage, Box, Button, Text } from '@island.is/island-ui/core'
 import { Loading } from '@island.is/judicial-system-web/src/shared-components'
-import { User, UserRole } from '@island.is/judicial-system/types'
+import { UserRole } from '@island.is/judicial-system/types'
+import type { User } from '@island.is/judicial-system/types'
 import * as Constants from '@island.is/judicial-system-web/src/utils/constants'
 import { useQuery } from '@apollo/client'
 import { UsersQuery } from '@island.is/judicial-system-web/src/utils/mutations'
