+++ apps/judicial-system/web/src/routes/Prosecutor/InvestigationRequest/HearingArrangements/HearingArrangements.tsx
@@ -16,9 +16,9 @@ import {
   CaseState,
   CaseTransition,
   NotificationType,
-  User,
   UserRole,
 } from '@island.is/judicial-system/types'
+import type { User } from '@island.is/judicial-system/types'
 import { useQuery } from '@apollo/client'
 import { CaseQuery } from '@island.is/judicial-system-web/graphql'
 import { UsersQuery } from '@island.is/judicial-system-web/src/utils/mutations'
