+++ apps/judicial-system/web/src/routes/Prosecutor/RestrictionRequest/StepTwo/StepTwo.tsx
@@ -8,9 +8,9 @@ import {
   CaseTransition,
   CaseType,
   NotificationType,
-  User,
   UserRole,
 } from '@island.is/judicial-system/types'
+import type { User } from '@island.is/judicial-system/types'
 import {
   CaseData,
   ProsecutorSubsections,
