+++ apps/judicial-system/web/src/shared-components/UserProvider/UserProvider.tsx
@@ -1,4 +1,4 @@
-import { User } from '@island.is/judicial-system/types'
+import type { User } from '@island.is/judicial-system/types'
 import { gql, useQuery } from '@apollo/client'
 import React, { createContext, useEffect, useState } from 'react'
 import { CSRF_COOKIE_NAME } from '@island.is/judicial-system/consts'
