+++ libs/judicial-system/auth/src/lib/decorators/graphql.user.decorator.ts
@@ -1,6 +1,6 @@
 import { createParamDecorator } from '@nestjs/common'
 
-import { User } from '@island.is/judicial-system/types'
+import type { User } from '@island.is/judicial-system/types'
 
 export const CurrentGraphQlUser = createParamDecorator(
   (data, { args: [_1, _2, { req }] }): User => req.user,
