+++ libs/judicial-system/auth/src/lib/guards/auth.guard.ts
@@ -1,7 +1,7 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common'
 import { AuthGuard } from '@nestjs/passport'
 
-import { User } from '@island.is/judicial-system/types'
+import type { User } from '@island.is/judicial-system/types'
 
 @Injectable()
 export class JwtAuthGuard extends AuthGuard('jwt') {
