+++ libs/judicial-system/auth/src/lib/guards/roles.guard.ts
@@ -1,7 +1,7 @@
 import { Injectable, CanActivate, ExecutionContext } from '@nestjs/common'
 import { Reflector } from '@nestjs/core'
 
-import { User } from '@island.is/judicial-system/types'
+import type { User } from '@island.is/judicial-system/types'
 
 import { RolesRule, RulesType } from '../auth.types'
