+++ libs/judicial-system/types/src/index.ts
@@ -3,11 +3,11 @@ export { Feature } from './lib/feature'
 export { InstitutionType } from './lib/institution'
 export type { Institution } from './lib/institution'
 
+export { UserRole } from './lib/user'
+export type { User, CreateUser, UpdateUser } from './lib/user'
+
 export * from './lib/types'
 export type {
-  User,
-  CreateUser,
-  UpdateUser,
   Case,
   Notification,
   CreateCase,
