+++ libs/judicial-system/types/src/lib/types.ts
@@ -1,46 +1,5 @@
-import { Institution } from './institution'
-
-export enum UserRole {
-  PROSECUTOR = 'PROSECUTOR',
-  REGISTRAR = 'REGISTRAR',
-  JUDGE = 'JUDGE',
-  ADMIN = 'ADMIN',
-}
-
-export interface User {
-  id: string
-  created: string
-  modified: string
-  nationalId: string
-  name: string
-  title: string
-  mobileNumber: string
-  email: string
-  role: UserRole
-  institution?: Institution
-  active: boolean
-}
-
-export interface CreateUser {
-  nationalId: string
-  name: string
-  title: string
-  mobileNumber: string
-  email: string
-  role: UserRole
-  institutionId: string
-  active: boolean
-}
-
-export interface UpdateUser {
-  name?: string
-  title?: string
-  mobileNumber?: string
-  email?: string
-  role?: UserRole
-  institutionId?: string
-  active?: boolean
-}
+import type { Institution } from './institution'
+import type { User } from './user'
 
 export enum CaseType {
   CUSTODY = 'CUSTODY',
